+++ R/zzz.R
@@ -1,0 +1,8 @@
+#' @import grid
+#' @importFrom R6 R6Class
+#' @importFrom grDevices dev.new devAskNewPage
+#' @importFrom rlang .data abort %||%
+#' @importFrom utils head packageDescription tail
+NULL
+
+hasName <- function(x, name) match(name, names(x), nomatch = 0L) > 0L